refactor: library detail view split from view controller
The only missing piece that I need to fix is that the main scroll
view is not responding.
