+++ PlantTracker/PlantTracker/Controllers/LibraryDetailViewController.swift
@@ -19,15 +19,7 @@ class LibraryDetailViewController: UIViewController, UIScrollViewDelegate {
     var plant: Plant!
     var plantsSaveDelegate: PlantsDelegate?
     
-    @IBOutlet var mainScrollView: UIScrollView!
-    @IBOutlet var headerView: UIView!
-    @IBOutlet var headerImageView: UIImageView!
-    var twicketSegementedControl: TwicketSegmentedControl!
-    @IBOutlet var informationView: UIView!
-    var floatyButton = Floaty()
-    var generalInfoTableView: UITableView!
-    var notesTextView: UITextView!
-    var linksTableView: UITableView!
+    var libraryDetailView: LibraryDetailView! = nil
     
     let keyboard = KeyboardObserver()
     
@@ -39,59 +31,63 @@ class LibraryDetailViewController: UIViewController, UIScrollViewDelegate {
     
     var assetTracker = AssetIndexIDTracker()
     
-    // height of header image
-    let headerImageHeight = 350
-    let minHeaderImageHeight = 100
-    
     var blurEffectView: UIVisualEffectView!
+    
+    
+//    override func loadView() {
+//
+//    }
+    
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        
+        libraryDetailView = LibraryDetailView()
+        view.addSubview(libraryDetailView)
+        libraryDetailView.snp.makeConstraints { make in make.edges.equalTo(view) }
+        
+        libraryDetailView.headerImage = getHeaderImage()
+        
+        libraryDetailView.setupView()
+        
+        libraryDetailView.mainScrollView.delegate = self
+        libraryDetailView.twicketSegementedControl.delegate = self
+        libraryDetailView.generalInfoTableView.delegate = self
+        libraryDetailView.generalInfoTableView.dataSource = self
+        libraryDetailView.linksTableView.delegate = self
+        libraryDetailView.linksTableView.dataSource = self
         
         title = plant.scientificName ?? plant.commonName ?? ""
         
         navigationItem.rightBarButtonItem = UIBarButtonItem(barButtonSystemItem: .edit, target: self, action: #selector(editDetailAction))
         
-        setupInformationTableViews()
-        setupDetailView()
         didSelect(0)
         hideKeyboardWhenTappedAround()
         setupKeyboardObserver()
     }
     
-    func scrollViewDidScroll(_ scrollView: UIScrollView) {
-        if scrollView == mainScrollView && startingYOffset == nil {
-            // initialize starting Y offset
-            startingYOffset = scrollView.contentOffset.y
-        }
-        
-        if scrollView == mainScrollView {
-            updateHeaderImage(offset: scrollView.contentOffset)
+    
+    func getHeaderImage() -> UIImage? {
+        if let imageID = plant.bestSingleImage() {
+            headerImageIsSet = true
+            return UIImage(contentsOfFile: getFilePathWith(id: imageID))
+        } else {
+            print("deafult header image")
+            return UIImage(named: "cactus")
         }
     }
     
-    func updateHeaderImage(offset: CGPoint) {
-        let scrollViewYDiff = startingYOffset! - offset.y
+    
+    func scrollViewDidScroll(_ scrollView: UIScrollView) {
+//        print("scrolling - x: \(scrollView.contentOffset.x), y: \(scrollView.contentOffset.y)")
         
-        // sticky header
-        let newHeight = max(CGFloat(headerImageHeight) + scrollViewYDiff, CGFloat(minHeaderImageHeight))
-        headerView.snp.remakeConstraints { (make) in
-            make.top.equalTo(view).offset(abs(startingYOffset!))
-            make.height.equalTo(newHeight)
-            make.trailing.equalTo(view)
-            make.leading.equalTo(view)
+        if scrollView == libraryDetailView.mainScrollView && libraryDetailView.startingYOffset == nil {
+            // initialize starting Y offset
+            libraryDetailView.startingYOffset = scrollView.contentOffset.y
         }
         
-        if scrollViewYDiff >= 0 {
-            // scrolling up
-            blurEffectView.alpha = 0.0
-        } else  if scrollViewYDiff < 0 {
-            // scrolling down
-            let frameHeight = CGFloat(headerView.frame.height)
-            let maxHeight = CGFloat(headerImageHeight)
-            let minHeight = CGFloat(minHeaderImageHeight)
-            let blurAlpha = (frameHeight - maxHeight) / (maxHeight - minHeight) * (0.0 - 0.85) + 0.0
-            blurEffectView.alpha = blurAlpha
+        if scrollView == libraryDetailView.mainScrollView {
+            libraryDetailView.updateHeaderImage(offset: scrollView.contentOffset)
         }
     }
     
@@ -104,6 +100,7 @@ class LibraryDetailViewController: UIViewController, UIScrollViewDelegate {
         alertController.addAction(UIAlertAction(title: "Remove plant from library", style: .destructive, handler: removeFromLibrary))
         alertController.addAction(UIAlertAction(title: "Cancel", style: .cancel))
         present(alertController, animated: true)
+        
     }
     
     
@@ -122,147 +119,6 @@ class LibraryDetailViewController: UIViewController, UIScrollViewDelegate {
         })
         present(alertControler, animated: true)
     }
-    
-    
-    func setupKeyboardObserver() {
-        keyboard.observe { [weak self] (event) in
-            
-            // for notes text view editing
-            if self?.notesTextView.isHidden == false {
-                switch event.type {
-                case .willHide:
-                    self?.notesTextView.contentInset = .zero
-                    
-                    if self?.notesTextView.text == "" {
-                        self?.notesTextView.text = "Notes"
-                        self?.notesTextView.textColor = .lightGray
-                    } else {
-                        self?.plant.notes = self?.notesTextView.text ?? ""
-                    }
-                    
-                case .willShow, .willChangeFrame:
-                    let keyboardScreenFrameEnd = event.keyboardFrameEnd
-                    let bottom = keyboardScreenFrameEnd.height - (self?.view.alignmentRectInsets.bottom)! + 8
-                    self?.notesTextView.contentInset.bottom = bottom
-                    
-                    if self?.notesTextView.text == "Notes" {
-                        self?.notesTextView.text = ""
-                        self?.notesTextView.textColor = .black
-                    }
-                    
-                    let scrollBottom = keyboardScreenFrameEnd.height
-                    self?.mainScrollView.setContentOffset(CGPoint(x: 0, y: scrollBottom), animated: true)
-                    
-                default:
-                    return
-                }
-            }
-        }
-        notesTextView.scrollIndicatorInsets = notesTextView.contentInset
-        notesTextView.scrollRangeToVisible(notesTextView.selectedRange)
-    }
-
-    
-    func setupDetailView() {
-        
-        // main scroll view
-        mainScrollView.delegate = self
-        mainScrollView.snp.makeConstraints { (make) in make.edges.equalTo(view) }
-        mainScrollView.showsVerticalScrollIndicator = false
-        mainScrollView.showsHorizontalScrollIndicator = false
-        
-        // header view
-        headerView.snp.makeConstraints { (make) in
-            make.top.equalTo(mainScrollView)
-            make.leading.equalTo(view)
-            make.trailing.equalTo(view)
-            make.height.equalTo(headerImageHeight)
-        }
-        
-        // header image
-        setHeaderImage()
-        headerImageView.contentMode = .scaleAspectFill
-        headerImageView.clipsToBounds = true
-        headerImageView.snp.makeConstraints { (make) in make.edges.equalTo(headerView) }
-        
-        // blur effect on header
-        blurEffectView = UIVisualEffectView(effect: UIBlurEffect(style: .dark))
-        blurEffectView.alpha = 0.0
-        headerView.addSubview(blurEffectView)
-        blurEffectView.snp.makeConstraints { (make) in
-            make.edges.equalTo(headerView)
-        }
-        
-        // twicket slider segmented control
-        twicketSegementedControl = TwicketSegmentedControl(frame: CGRect(x: 5, y: 0, width: view.frame.height - 10, height: 40))
-        twicketSegementedControl.setSegmentItems(["Information", "Notes", "Links"])
-        twicketSegementedControl.delegate = self
-        view.addSubview(twicketSegementedControl)
-        twicketSegementedControl.snp.makeConstraints { (make) in
-            make.top.equalTo(headerView.snp.bottom)
-            make.leading.equalTo(view)
-            make.trailing.equalTo(view)
-            make.height.equalTo(35)
-        }
-        
-        // floatly button
-        setUpFloatlyButton()
-        let padding = 20
-        floatyButton.paddingX = CGFloat(padding)
-        floatyButton.paddingY = CGFloat(padding)
-        let floatlyFrame = floatyButton.frame
-        headerView.addSubview(floatyButton)
-        floatyButton.snp.makeConstraints { make in
-            make.bottom.equalTo(headerView.snp.bottom).inset(padding)
-            make.right.equalTo(headerView.snp.right).inset(padding)
-            make.width.equalTo(floatlyFrame.width)
-            make.height.equalTo(floatlyFrame.height)
-        }
-        
-        // information view (below segmented control)
-        informationView.snp.makeConstraints { (make) in
-            make.top.equalTo(twicketSegementedControl.snp.bottom).offset(5)
-            make.leading.equalTo(view)
-            make.trailing.equalTo(view)
-            make.bottom.equalTo(view.snp.bottom)
-        }
-        
-        // information subviews
-        notesTextView = UITextView()
-        informationView.addSubview(generalInfoTableView)
-        informationView.addSubview(notesTextView)
-        informationView.addSubview(linksTableView)
-        generalInfoTableView.snp.makeConstraints { (make) in make.edges.equalTo(informationView)}
-        notesTextView.snp.makeConstraints { (make) in make.edges.equalTo(informationView).offset(8) }
-        linksTableView.snp.makeConstraints { (make) in make.edges.equalTo(informationView) }
-        
-        // set up notes text view
-        notesTextView.textAlignment = .left
-        notesTextView.returnKeyType = .default
-        notesTextView.font = UIFont.systemFont(ofSize: 17)
-        if plant.notes.count > 0 {
-            notesTextView.text = plant.notes
-            notesTextView.textColor = .black
-        } else {
-            notesTextView.text = "Notes"
-            notesTextView.textColor = .lightGray
-        }
-        
-        
-        // initalize content height of main scroll view
-        let contentHeight = CGFloat(headerImageHeight) + view.frame.height - CGFloat(minHeaderImageHeight)
-        mainScrollView.contentSize = CGSize(width: view.frame.width, height: contentHeight)
-    }
-    
-    func setHeaderImage() {
-        if let imageID = plant.bestSingleImage() {
-            headerImageView.image = UIImage(contentsOfFile: getFilePathWith(id: imageID))
-            headerImageIsSet = true
-        } else {
-            print("deafult header image")
-            headerImageView.image = UIImage(named: "cactus")
-        }
-    }
 }
 
 
@@ -270,25 +126,11 @@ class LibraryDetailViewController: UIViewController, UIScrollViewDelegate {
 
 extension LibraryDetailViewController: UITableViewDelegate, UITableViewDataSource {
     
-    func setupInformationTableViews() {
-        generalInfoTableView = UITableView.init(frame: CGRect.zero, style: .plain)
-        generalInfoTableView.delegate = self
-        generalInfoTableView.dataSource = self
-        generalInfoTableView.register(GeneralInformtationTableViewCell.self, forCellReuseIdentifier: "generalInfoCell")
-        generalInfoTableView.allowsSelection = false
-        
-        linksTableView = UITableView.init(frame: CGRect.zero, style: .plain)
-        linksTableView.delegate = self
-        linksTableView.dataSource = self
-        linksTableView.register(UITableViewCell.self, forCellReuseIdentifier: "linksCell")
-    }
-    
-    
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         switch tableView {
-        case generalInfoTableView:
+        case libraryDetailView.generalInfoTableView:
             return 7
-        case linksTableView:
+        case libraryDetailView.linksTableView:
             return 3
         default:
             fatalError("Unforeseen table view requesting number of cells")
@@ -298,7 +140,7 @@ extension LibraryDetailViewController: UITableViewDelegate, UITableViewDataSourc
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         switch tableView {
-        case generalInfoTableView:
+        case libraryDetailView.generalInfoTableView:
             let cell = tableView.dequeueReusableCell(withIdentifier: "generalInfoCell", for: indexPath)
             var main: String?
             var detail: String?
@@ -332,7 +174,7 @@ extension LibraryDetailViewController: UITableViewDelegate, UITableViewDataSourc
             cell.textLabel?.text = main
             cell.detailTextLabel?.text = detail
             return cell
-        case linksTableView:
+        case libraryDetailView.linksTableView:
             let cell = tableView.dequeueReusableCell(withIdentifier: "linksCell", for: indexPath)
             cell.textLabel?.text = "TEST LINKS - row\(indexPath.row)"
             return cell
@@ -344,9 +186,9 @@ extension LibraryDetailViewController: UITableViewDelegate, UITableViewDataSourc
     
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
         switch tableView {
-        case generalInfoTableView:
+        case libraryDetailView.generalInfoTableView:
             return 50
-        case linksTableView:
+        case libraryDetailView.linksTableView:
             return 75
         default:
             return 44
@@ -355,7 +197,7 @@ extension LibraryDetailViewController: UITableViewDelegate, UITableViewDataSourc
     
     func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
         switch tableView {
-        case generalInfoTableView:
+        case libraryDetailView.generalInfoTableView:
             return true
         default:
             return false
@@ -366,26 +208,26 @@ extension LibraryDetailViewController: UITableViewDelegate, UITableViewDataSourc
 
 
 
-// add extension for Twicket segmented control
+// MARK: TwicketSegmentedControlDelegate
 extension LibraryDetailViewController: TwicketSegmentedControlDelegate {
     func didSelect(_ segmentIndex: Int) {
-        switch twicketSegementedControl.selectedSegmentIndex {
+        switch libraryDetailView.twicketSegementedControl.selectedSegmentIndex {
         case 0:
             // "Information"
-            generalInfoTableView.isHidden = false
-            notesTextView.isHidden = true
-            linksTableView.isHidden = true
+            libraryDetailView.generalInfoTableView.isHidden = false
+            libraryDetailView.notesTextView.isHidden = true
+            libraryDetailView.linksTableView.isHidden = true
             dismissKeyboard()
         case 1:
             // "Notes"
-            generalInfoTableView.isHidden = true
-            notesTextView.isHidden = false
-            linksTableView.isHidden = true
+            libraryDetailView.generalInfoTableView.isHidden = true
+            libraryDetailView.notesTextView.isHidden = false
+            libraryDetailView.linksTableView.isHidden = true
         case 2:
             // "Links"
-            generalInfoTableView.isHidden = true
-            notesTextView.isHidden = true
-            linksTableView.isHidden = false
+            libraryDetailView.generalInfoTableView.isHidden = true
+            libraryDetailView.notesTextView.isHidden = true
+            libraryDetailView.linksTableView.isHidden = false
             dismissKeyboard()
         default:
             break
@@ -395,10 +237,51 @@ extension LibraryDetailViewController: TwicketSegmentedControlDelegate {
 
 
 
-// to dismiss keyboard with taps anywhere else in view
-extension UIViewController {
+// MARK: Keyboard
+
+extension LibraryDetailViewController {
+    
+    func setupKeyboardObserver() {
+        keyboard.observe { [weak self] (event) in
+            
+            // for notes text view editing
+            if self?.libraryDetailView.notesTextView.isHidden == false {
+                switch event.type {
+                case .willHide:
+                    self?.libraryDetailView.notesTextView.contentInset = .zero
+                    
+                    if self?.libraryDetailView.notesTextView.text == "" {
+                        self?.libraryDetailView.notesTextView.text = "Notes"
+                        self?.libraryDetailView.notesTextView.textColor = .lightGray
+                    } else {
+                        self?.plant.notes = self?.libraryDetailView.notesTextView.text ?? ""
+                    }
+                    
+                case .willShow, .willChangeFrame:
+                    let keyboardScreenFrameEnd = event.keyboardFrameEnd
+                    let bottom = keyboardScreenFrameEnd.height - (self?.view.alignmentRectInsets.bottom)! + 8
+                    self?.libraryDetailView.notesTextView.contentInset.bottom = bottom
+                    
+                    if self?.libraryDetailView.notesTextView.text == "Notes" {
+                        self?.libraryDetailView.notesTextView.text = ""
+                        self?.libraryDetailView.notesTextView.textColor = .black
+                    }
+                    
+                    let scrollBottom = keyboardScreenFrameEnd.height
+                    self?.libraryDetailView.mainScrollView.setContentOffset(CGPoint(x: 0, y: scrollBottom), animated: true)
+                    
+                default:
+                    return
+                }
+            }
+        }
+        libraryDetailView.notesTextView.scrollIndicatorInsets = libraryDetailView.notesTextView.contentInset
+        libraryDetailView.notesTextView.scrollRangeToVisible(libraryDetailView.notesTextView.selectedRange)
+    }
+    
+    
     func hideKeyboardWhenTappedAround() {
-        let tap = UITapGestureRecognizer(target: self, action: #selector(UIViewController.dismissKeyboard))
+        let tap = UITapGestureRecognizer(target: self, action: #selector(LibraryDetailViewController.dismissKeyboard))
         tap.cancelsTouchesInView = false
         view.addGestureRecognizer(tap)
     }
@@ -415,6 +298,7 @@ extension LibraryDetailViewController: AssetsPickerViewControllerDelegate, UINav
     func assetsPickerCannotAccessPhotoLibrary(controller: AssetsPickerViewController) {
         print("Need permission to access photo library.")
     }
+    
     
     @objc func addImages(_ alert: UIAlertAction) {
         
@@ -434,6 +318,7 @@ extension LibraryDetailViewController: AssetsPickerViewControllerDelegate, UINav
         print("opening image picker")
         present(imagePicker, animated: true)
     }
+    
     
     func assetsPicker(controller: AssetsPickerViewController, didSelect asset: PHAsset, at indexPath: IndexPath) {
         let imageManager = PHImageManager.default()
@@ -461,6 +346,7 @@ extension LibraryDetailViewController: AssetsPickerViewControllerDelegate, UINav
         print("saving request ID '\(requestIndex)' to index path '\(indexPath.item)'")
     }
     
+    
     func assetsPicker(controller: AssetsPickerViewController, didDeselect asset: PHAsset, at indexPath: IndexPath) {
         if let uuid = assetTracker.uuidFrom(indexPathItem: indexPath.item) {
             print("deleting image uuid '\(uuid)' at index path '\(indexPath.item)'")
@@ -470,31 +356,8 @@ extension LibraryDetailViewController: AssetsPickerViewControllerDelegate, UINav
         }
     }
     
+    
     func assetsPicker(controller: AssetsPickerViewController, selected assets: [PHAsset]) {
-//        let imageManager = PHImageManager.default()
-//        let imageOptions = PHImageRequestOptions()
-//
-//        let defaults = UserDefaults.standard
-//        switch defaults.string(forKey: "image quality") {
-//        case "high":
-//            imageOptions.deliveryMode = .highQualityFormat
-//        case "medium":
-//            imageOptions.deliveryMode = .opportunistic
-//        case "low":
-//            imageOptions.deliveryMode = .fastFormat
-//        default:
-//            imageOptions.deliveryMode = .highQualityFormat
-//            print("value not entered for \"Image Quality\" setting.")
-//        }
-//        imageOptions.version = .current
-//        imageOptions.isSynchronous = false
-//        imageOptions.resizeMode = .exact
-//
-//        print("selected \(assets.count) images")
-//        for asset in assets {
-//            let assetSize = CGSize(width: Double(asset.pixelWidth), height: Double(asset.pixelHeight))
-//            imageManager.requestImage(for: asset, targetSize: assetSize, contentMode: .aspectFit, options: imageOptions, resultHandler: addImageToPlant)
-//        }
         for index in assetTracker.didNotDeleteAtRequestIndex {
             if let uuid = assetTracker.uuidFrom(indexPathItem: index) {
                 print("deleting image uuid '\(uuid)' at index path '\(index)'")
@@ -503,8 +366,9 @@ extension LibraryDetailViewController: AssetsPickerViewControllerDelegate, UINav
         }
         assetTracker.reset()
         if let delegate = plantsSaveDelegate { delegate.savePlants() }
-        setHeaderImage()
+        libraryDetailView.headerImage = getHeaderImage()
     }
+    
     
     func assetsPickerDidCancel(controller: AssetsPickerViewController) {
         print("user canceled asset getting")
@@ -539,11 +403,6 @@ extension LibraryDetailViewController: AssetsPickerViewControllerDelegate, UINav
                     if let info = info { print(info) }
                     print("-----------------")
                 }
-//                if !(self!.headerImageIsSet) {
-//                    DispatchQueue.main.async { self?.setHeaderImage() }
-//                }
-                
-//                if let delegate = self?.plantsSaveDelegate { delegate.savePlants() }
             }
         }
     }
@@ -552,7 +411,8 @@ extension LibraryDetailViewController: AssetsPickerViewControllerDelegate, UINav
 
 
 
-// handle segues
+// MARK: segue
+
 extension LibraryDetailViewController {
     
     // seque into image collection view
@@ -562,45 +422,5 @@ extension LibraryDetailViewController {
             vc.imageIDs = plant.images
             vc.title = self.title
         }
-    }
-}
-
-
-// floaty button
-extension LibraryDetailViewController {
-    func setUpFloatlyButton() {
-        
-        // set self to delegate (maybe)
-//        floatyButton.fabDelegate = self
-        
-        // stlying
-        floatyButton.relativeToSafeArea = false
-        floatyButton.sticky = true
-        floatyButton.hasShadow = true
-        floatyButton.buttonShadowColor = .darkGray
-        floatyButton.buttonColor = .gray
-        
-        floatyButton.autoCloseOnTap = true
-        floatyButton.isUserInteractionEnabled = true
-        floatyButton.isHidden = false
-        floatyButton.openAnimationType = .slideUp
-        floatyButton.animationSpeed = 1.0
-        
-        // item 1: add photos
-        let addPhotosItem = FloatyItem()
-        addPhotosItem.title = "Add photos"
-        addPhotosItem.icon = UIImage(named: "cameraIconBW")
-        addPhotosItem.titleColor = .white
-        addPhotosItem.buttonColor = .lightGray
-//        addPhotosItem.handler = addImages()
-        floatyButton.addItem(item: addPhotosItem)
-        
-        // item 2: view all photos
-        let viewPhotosItem = FloatyItem()
-        viewPhotosItem.title = "View photos"
-        viewPhotosItem.titleColor = .white
-        viewPhotosItem.buttonColor = .lightGray
-        viewPhotosItem.icon = UIImage(named: "albumIconBW")
-        floatyButton.addItem(item: viewPhotosItem)
     }
 }
